part5: Remove usages of `GrowableIOBuffer::StartOfBuffer`
This CL removes usages of the `GrowableIOBuffer::StartOfBuffer` method
and replaces unchecked pointers with `span`. Additionally
`HttpCache::ParseResponseInfo` now consumes a `span` instead of a
ptr+size pair.

Bug: 329476354
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5706188
Commit-Queue: kelsen liu <kelsenliu21@gmail.com>
Cr-Commit-Position: refs/heads/main@{#1329299}

+++ net/http/http_stream_parser.h
@@ -208,7 +208,7 @@ class NET_EXPORT_PRIVATE HttpStreamParser {
 
   // Parse the headers into response_.  Returns OK on success or a net::Error on
   // failure.
-  int ParseResponseHeaders(int end_of_header_offset);
+  int ParseResponseHeaders(size_t end_of_header_offset);
 
   // Examine the parsed headers to try to determine the response body size.
   void CalculateResponseBodySize();
